Move data_hyperx to libratbag-data.h
