+++ src/libratbag-data.h
@@ -193,3 +193,20 @@ ratbag_device_data_asus_is_wireless(const struct ratbag_device_data *data);
  */
 uint32_t
 ratbag_device_data_asus_get_quirks(const struct ratbag_device_data *data);
+
+struct data_hyperx {
+	int profile_count;
+	int button_count;
+	int led_count;
+	size_t nrates;
+	unsigned int *rates;
+	int dpi_count;
+	int is_wireless;
+	struct dpi_range *dpi_range;
+};
+
+/**
+ * @return HyperX device data
+ */
+const struct data_hyperx *
+ratbag_device_data_hyperx_get_data(const struct ratbag_device_data *data);
